Add support for kubeconfig plugin vairable

+++ plugin.sh
@@ -18,32 +18,43 @@ if [ -z "$PLUGIN_FOLDERPATH" ]; then
     exit 1
 fi
 
-## Auto load cluster config with a couple steps
-## Step1: Login to az account
-## Step2: Load the cluster config to connect the connection
-echo ">>> Adding AKS cluster as config for connection <<<"
+echo ">>> Adding AKS cluster config for connection <<<"
 
-echo ">>> Signing into Azure <<<"
-az login --service-principal -u ${PLUGIN_AZURE_APPID} -p ${PLUGIN_AZURE_PASSWORD} --tenant ${PLUGIN_AZURE_TENANT} || exit 1
+## IF PLUGIN_KUBECONFIG is not passed try connecting to the cluster with `az login...`
+## otherwise add kubeconfig in the $HOME dierctory to connect with aks cluster
+if [ -z "$PLUGIN_KUBECONFIG" ]; then
+    ## Auto load cluster config with a couple steps
+    ## Step1: Login to az account
+    ## Step2: Load the cluster config to connect the connection
 
-echo ">>> Adding Cluster $HOME/.kube/config <<<"
-az aks get-credentials --name ${PLUGIN_CLUSTER} --resource-group ${PLUGIN_CLUSTER_RG} || exit 1
+    echo ">>> Signing into Azure <<<"
+    az login --service-principal -u ${PLUGIN_AZURE_APPID} -p ${PLUGIN_AZURE_PASSWORD} --tenant ${PLUGIN_AZURE_TENANT} || exit 1
 
-echo ">>> Checking for the deployment operation to be performed. It could be DB migration job or k8s resource deployment like: deployment or namespace <<<"
+    echo ">>> Adding Cluster $HOME/.kube/config <<<"
+    az aks get-credentials --name ${PLUGIN_CLUSTER} --resource-group ${PLUGIN_CLUSTER_RG} || exit 1
 
-## Delete the existing `migration` job if it exists.
-## New migrations cannot be deployed without deleting the existing one.
-PLUGIN_MIGRATION_JOB="${PLUGIN_MIGRATION_JOB:-false}"
-if [ $PLUGIN_MIGRATION_JOB == true ]
-    then
-    PLUGIN_NAMESPACE="${PLUGIN_NAMESPACE:-default}"
-    if [ PLUGIN_NAMESPACE != "default" ]
+    echo ">>> Checking for the deployment operation to be performed. It could be DB migration job or k8s resource deployment like: deployment or namespace <<<"
+
+    ## Delete the existing `migration` job if it exists.
+    ## New migrations cannot be deployed without deleting the existing one.
+    PLUGIN_MIGRATION_JOB="${PLUGIN_MIGRATION_JOB:-false}"
+    if [ $PLUGIN_MIGRATION_JOB == true ]
         then
-        echo ">>> Deleting the existing DB migration Job resource: ${PLUGIN_JOBNAME} in Namespace: ${PLUGIN_NAMESPACE}. <<<"
-        kubectl delete -n ${PLUGIN_NAMESPACE} job/${PLUGIN_JOBNAME} || true
-    else
-        echo ">>> Error: No namespace defined for the migration job. <<<"
+        PLUGIN_NAMESPACE="${PLUGIN_NAMESPACE:-default}"
+        if [ PLUGIN_NAMESPACE != "default" ]
+            then
+            echo ">>> Deleting the existing DB migration Job resource: ${PLUGIN_JOBNAME} in Namespace: ${PLUGIN_NAMESPACE}. <<<"
+            kubectl delete -n ${PLUGIN_NAMESPACE} job/${PLUGIN_JOBNAME} || true
+        else
+            echo ">>> Error: No namespace defined for the migration job. <<<"
+        fi
     fi
+else
+    echo ">>> Setting kubeconfig to access the k8s cluster <<<"
+    [ -d $HOME/.kube ] || mkdir $HOME/.kube
+    echo "# Plugin PLUGIN_KUBECONFIG available" >&2
+    echo "$PLUGIN_KUBECONFIG" > $HOME/.kube/config
+    unset PLUGIN_KUBECONFIG
 fi
 
 cd "${PLUGIN_FOLDERPATH}"
